playing around with backing up files and such

+++ quick-replace.sh
@@ -1,15 +1,26 @@
 #!/bin/bash
 
-# backup the given document
-~/gitrepos/scripts/backup.sh $1;
+E_NO_FILE_ERR=64
 
-# make a temporary file
-NEWFILE=temp_\$1;
+if [ $# -eq 1 ]
+then
+    # backup the given document
+    ~/gitrepos/scripts/backup.sh $1;
 
-while read LINE; do
-    echo $LINE >> $NEWFILE
-done
+    # make a temporary file
+    NEWFILE=temp_\$1;
 
-# replace $1 and remove temporary file
-cat $NEWFILE > $1;
-rm $NEWFILE;
+    while read LINE; do
+        echo $LINE >> $NEWFILE
+    done
+
+    # replace $1 and remove temporary file
+    cat $NEWFILE > $1;
+    rm $NEWFILE;
+
+else
+    echo "first arg must be a regular file that already exists"
+    exit $E_NO_FILE_ERR
+fi
+
+exit 0
